GSIN-backend/MemoryClusterNetworks/requirements.txt
numpy
pandas
scikit-learn
faiss-cpu
sentence-transformers
matplotlib
